requirements.txt
requests==2.32.4
python-dotenv==1.1.1
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.26
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
pydantic==2.11.7
nest-asyncio==1.6.0
langchain-groq==0.3.4
langchain-sambanova==1.0.0
langchain-ollama==0.3.3
playwright==1.53.0
crawl4ai==0.7.7
beautifulsoup4==4.13.4
httpx==0.28.1
langgraph==0.5.0
chromadb==1.3.5
faiss-cpu==1.11.0